backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
openai>=1.54.0
pydantic>=2.10.0
python-dotenv>=1.0.0
backoff>=2.2.1
pypdf>=5.0.0
psycopg[binary]>=3.2.0
requests>=2.31.0

